cat/format: internally invoke hashing to ensure deep readiness
